document_ai_warehouse/document_ai_warehouse_processing_python/requirements.txt
google-cloud-contentwarehouse==0.2.1
google-cloud-core==2.3.2
google-cloud-documentai==2.7.0
google-cloud-storage==2.7.0
googleapis-common-protos==1.58.0
tabulate==0.9.0
